settings panel (#344)
* Add settings panel to NVDA Remote configuration to mirror existing options dialog

Creates a new RemoteSettingsPanel class that adds remote connection settings
to the NVDA Settings dialog when not running in secure mode. Includes options
for auto-connect, connection type, server details, and sound preferences.
Add config loading/saving and UI event handling for the Remote settings panel.

* refactor: Extract autoconnect validation logic into isValid() method

* feat: remove options dialog and menu item

This change removes the old options dialog functionality and its associated tools > remote > options menu item in favor of the new remote settings panel:
* Removed displayOptionsInterface method from GlobalPlugin
* Removed options menu item from RemoteMenu
* remove options dialog class

* refactor: replace port TextCtrl with SpinCtrl for validated input

This change replaces the port text input with a SpinCtrl widget that enforces valid port numbers (1-65535), improving input validation.

* Updated error dialog strings to indicate 'remote' is the source of the error.
