+++ addon/globalPlugins/remoteClient/menu.py
@@ -11,7 +11,6 @@ class RemoteMenu(wx.Menu):
 	muteItem: wx.MenuItem
 	pushClipboardItem: wx.MenuItem
 	copyLinkItem: wx.MenuItem
-	optionsItem: wx.MenuItem
 	sendCtrlAltDelItem: wx.MenuItem
 	remoteItem: wx.MenuItem
 
@@ -38,9 +37,6 @@ class RemoteMenu(wx.Menu):
 		self.copyLinkItem: wx.MenuItem = self.Append(wx.ID_ANY, _("Copy &link"), _("Copy a link to the remote session"))
 		self.copyLinkItem.Enable(False)
 		gui.mainFrame.sysTrayIcon.Bind(wx.EVT_MENU, self.onCopyLinkItem, self.copyLinkItem)
-		# Translators: Menu item in NvDA Remote submenu to open add-on options.
-		self.optionsItem: wx.MenuItem = self.Append(wx.ID_ANY, _("&Options..."), _("Options"))
-		gui.mainFrame.sysTrayIcon.Bind(wx.EVT_MENU, self.onOptionsItem, self.optionsItem)
 		# Translators: Menu item in NVDA Remote submenu to send Control+Alt+Delete to the remote computer.
 		self.sendCtrlAltDelItem: wx.MenuItem = self.Append(wx.ID_ANY, _("Send Ctrl+Alt+Del"), _("Send Ctrl+Alt+Del"))
 		gui.mainFrame.sysTrayIcon.Bind(wx.EVT_MENU, self.onSendCtrlAltDel, self.sendCtrlAltDelItem)
@@ -64,9 +60,6 @@ class RemoteMenu(wx.Menu):
 		self.Remove(self.copyLinkItem.Id)
 		self.copyLinkItem.Destroy()
 		self.copyLinkItem = None
-		self.Remove(self.optionsItem.Id)
-		self.optionsItem.Destroy()
-		self.optionsItem=None
 		self.Remove(self.sendCtrlAltDelItem.Id)
 		self.sendCtrlAltDelItem.Destroy()
 		self.sendCtrlAltDelItem=None
@@ -94,10 +87,6 @@ class RemoteMenu(wx.Menu):
 	def onCopyLinkItem(self, evt: wx.CommandEvent) -> None:
 		evt.Skip()
 		self.client.copyLink()
-
-	def onOptionsItem(self, evt: wx.CommandEvent) -> None:
-		evt.Skip()
-		self.client.displayOptionsInterface()
 
 	def onSendCtrlAltDel(self, evt: wx.CommandEvent) -> None:
 		evt.Skip()
